test_prompt_templates.py
#!/usr/bin/env python3
"""
Test BigAcademy Generic Prompt Template System
"""

import sys
sys.path.append('/Users/franckbirba/DEV/TEST-CREWAI/bigacademy')

from bigacademy.core.agent_profiles import AgentProfileManager
from bigacademy.core.graph_db import GraphDB
from bigacademy.generators.prompt_templates import PromptTemplateManager
from bigacademy.extractors.base_extractor import KnowledgeChunk
from pathlib import Path

def test_template_loading():
    """Test loading template configurations"""
    print("🎯 Testing Template Loading")
    print("=" * 50)
    
    templates_dir = Path("/Users/franckbirba/DEV/TEST-CREWAI/bigacademy/configs/templates")
    template_manager = PromptTemplateManager(templates_dir)
    
    available_templates = template_manager.get_available_templates()
    print(f"📋 Available templates: {available_templates}")
    
    for template_type in available_templates:
        info = template_manager.get_template_info(template_type)
        print(f"\n📝 Template: {template_type}")
        print(f"   Description: {info['description']}")
        print(f"   Suitable for: {info['suitable_for']}")
        print(f"   Content types: {info['content_types']}")
        print(f"   Variables: {len(info['variables'])} required")
    
    return template_manager

def test_agent_template_matching():
    """Test matching templates to agent profiles"""
    print("\n🤖 Testing Agent Template Matching")
    print("=" * 50)
    
    # Load agents
    agent_manager = AgentProfileManager(Path("/Users/franckbirba/DEV/TEST-CREWAI/bigacademy/configs/agents"))
    template_manager = PromptTemplateManager(Path("/Users/franckbirba/DEV/TEST-CREWAI/bigacademy/configs/templates"))
    
    agents = ["solution_architect", "python_developer"]
    
    for agent_name in agents:
        agent_profile = agent_manager.get_profile(agent_name)
        if not agent_profile:
            print(f"❌ Could not load {agent_name}")
            continue
        
        print(f"\n🤖 Agent: {agent_profile.name}")
        print(f"   Role: {agent_profile.role.title}")
        print(f"   Focus Areas: {', '.join(agent_profile.focus_areas[:3])}")
        
        # Test different content types
        content_types = ["code", "documentation", "configuration"]
        
        for content_type in content_types:
            suitable_templates = template_manager.get_suitable_templates(agent_profile, content_type)
            print(f"   📄 {content_type}: {suitable_templates}")
    
    return agent_manager, template_manager

def test_prompt_generation():
    """Test generating actual prompts from templates"""
    print("\n🎨 Testing Prompt Generation")
    print("=" * 50)
    
    # Load components
    agent_manager = AgentProfileManager(Path("/Users/franckbirba/DEV/TEST-CREWAI/bigacademy/configs/agents"))
    template_manager = PromptTemplateManager(Path("/Users/franckbirba/DEV/TEST-CREWAI/bigacademy/configs/templates"))
    
    # Get architect profile
    architect = agent_manager.get_profile("solution_architect")
    if not architect:
        print("❌ Could not load architect profile")
        return
    
    # Create mock knowledge chunk
    mock_chunk = KnowledgeChunk(
        content="""from fastapi import FastAPI, Depends
from fastapi.security import HTTPBearer

app = FastAPI()
security = HTTPBearer()

@app.get("/protected")
def protected_route(token: str = Depends(security)):
    return {"message": "Access granted", "token": token}""",
        source_path="auth/main.py",
        file_type=".py",
        language="python",
        size_tokens=150,
        relevance_score=0.85,
        metadata={"framework": "fastapi"}
    )
    
    source_info = {
        "url": "https://github.com/fastapi-users/fastapi-users",
        "type": "github_repository"
    }
    
    # Test different template types
    template_types = ["question_answer", "code_review", "implementation_task"]
    
    for template_type in template_types:
        print(f"\n📝 Generating prompt with template: {template_type}")
        print("-" * 60)
        
        try:
            prompt = template_manager.generate_prompt(
                template_type=template_type,
                agent_profile=architect,
                knowledge_chunk=mock_chunk,
                source_info=source_info,
                question_type="security implementation"
            )
            
            # Show first 300 characters of prompt
            preview = prompt[:300] + "..." if len(prompt) > 300 else prompt
            print(preview)
            print(f"\n✅ Full prompt length: {len(prompt)} characters")
            
        except Exception as e:
            print(f"❌ Error generating prompt: {e}")

def test_with_real_knowledge():
    """Test prompt generation with real extracted knowledge"""
    print("\n🧠 Testing with Real Knowledge from Graph DB")
    print("=" * 50)
    
    # Load components
    agent_manager = AgentProfileManager(Path("/Users/franckbirba/DEV/TEST-CREWAI/bigacademy/configs/agents"))
    template_manager = PromptTemplateManager(Path("/Users/franckbirba/DEV/TEST-CREWAI/bigacademy/configs/templates"))
    
    # Load from real graph database
    db_path = Path("test_data/fastapi_minimal_architect.db")
    if not db_path.exists():
        print(f"⚠️  Database not found: {db_path}")
        return
    
    db = GraphDB(db_path)
    architect = agent_manager.get_profile("solution_architect")
    
    if not architect:
        print("❌ Could not load architect profile")
        return
    
    # Get real knowledge chunks
    agents = db.find_nodes("Agent", {"name": "solution_architect"})
    if not agents:
        print("❌ No agent found in database")
        return
    
    agent_id = agents[0].id
    learn_relationships = db.get_relationships(agent_id, "LEARNS_FROM", "outgoing")
    
    print(f"📚 Found {len(learn_relationships)} knowledge chunks")
    
    # Test with top 3 chunks
    for i, rel in enumerate(learn_relationships[:3]):
        chunk_node = db.get_node(rel.target_id)
        if not chunk_node:
            continue
        
        # Create knowledge chunk object
        real_chunk = KnowledgeChunk(
            content=chunk_node.properties.get('content', ''),
            source_path=chunk_node.properties.get('source_path', ''),
            file_type=chunk_node.properties.get('file_type', ''),
            language=chunk_node.properties.get('language', 'text'),
            size_tokens=chunk_node.properties.get('size_tokens', 0),
            relevance_score=chunk_node.properties.get('relevance_score', 0),
            metadata=chunk_node.properties.get('metadata', {})
        )
        
        source_info = {
            "url": "https://github.com/fastapi-users/fastapi-users",
            "type": "github_repository"
        }
        
        print(f"\n📝 Knowledge Chunk {i+1}: {real_chunk.source_path}")
        print(f"   Relevance: {real_chunk.relevance_score:.3f}, Tokens: {real_chunk.size_tokens}")
        
        # Generate question-answer prompt
        try:
            prompt = template_manager.generate_prompt(
                template_type="question_answer",
                agent_profile=architect,
                knowledge_chunk=real_chunk,
                source_info=source_info
            )
            
            print(f"   ✅ Generated prompt: {len(prompt)} characters")
            
            # Show sample of the generated prompt
            lines = prompt.split('\n')
            print(f"   📋 Sample: {lines[0] if lines else 'Empty prompt'}")
            
        except Exception as e:
            print(f"   ❌ Error: {e}")
    
    db.close()

def main():
    """Run all prompt template tests"""
    print("🎯 BigAcademy Generic Prompt Template Test Suite")
    print("=" * 70)
    
    try:
        # Test 1: Template loading
        template_manager = test_template_loading()
        
        # Test 2: Agent template matching
        agent_manager, template_manager = test_agent_template_matching()
        
        # Test 3: Prompt generation with mock data
        test_prompt_generation()
        
        # Test 4: Real knowledge integration
        test_with_real_knowledge()
        
        print("\n🎉 Prompt template test suite completed!")
        print("\n💡 Key Results:")
        print("   ✅ Generic template system working")
        print("   ✅ Agent-agnostic template matching")
        print("   ✅ Dynamic prompt generation")
        print("   ✅ Real knowledge integration")
        print("   🚀 Ready for dataset generation!")
        
    except KeyboardInterrupt:
        print("\n⏹️  Test interrupted by user")
    except Exception as e:
        print(f"\n❌ Test suite failed: {e}")
        import traceback
        traceback.print_exc()

if __name__ == "__main__":
    main()