requirements.txt
# Core dependencies
pydantic>=2.0.0
pyyaml>=6.0
networkx>=3.0
requests>=2.28.0

# Database
sqlite3  # Built into Python

# Knowledge extraction
gpt-repository-loader>=0.1.0

# Dataset generation (when available)
# distilabel>=1.0.0  # Uncomment when implementing LLM integration

# Development and testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Optional: for visualization
matplotlib>=3.5.0
plotly>=5.0.0